Experimented wiring up the GraphicsCore and the Client, using some hack for preserving the current version of rendering. Need to refactor when we have access to the demo computers.
Refactored EffectCompositor and CompositorAnalysis classes.

Introduced editor mode and game mode.
Basic idea is that in the editor mode, we are using a trackball combined
with a modified fps camera for artist's convenience.
In the game mode, the camera should be switched to the real fps camera.
Also in the game mode, all gui's (except in game guis) are disabled.

You can enter the game mode by pressing "Run Game" on the editor bar.
You can exit the game mode by pressing "ESC" button. (need to change this
button when we demo, or we are using cheats...)

Currently, game mode fps camera is not merged with Philip's
implementation.