Big update! 0) Fixed the existing PBR shading model (made some mistakes previously)
1) Completed both diffuse and specular components of physically based IBL. Now objects are shaded by their environment map. Reduced the size of reflection cubemap to 128x128, speeding up the precompute stage.

2) Created a defualtAsset folder and an assets folder. The former is
used as minimum setup for running the editor. The latter is used to store
game assets loaded in the editor such as geom objects, materials, and
world files. The assets folder will not be uploaded to the github.
Excluded by .gitignore. Also, defaultAssets folder is not ignored for any
file in it.

3) Added first person camera manipulator in the editor.

4) CubemapUtil can take hdr files as input

5) The editor's camera is now at (0, 0, 10), looking down the y-axis.

6) Created a fake infinite plane for the ground with textures applied.

7) Fixed camera manipulator not correctly preserve original position if
unlocked. (Due to the misunderstanding of cv->getLookVectorLocal(). The
real lookAt vector is  cv->getEyeLocal())cv->getLookVectorLocal();

8) Fixed detachManipulator() won't deselect the corresponding manipulator
in the GUI.

9) Editor usability improved:
	1) Can only change geom shape when cam is locked
	2) By pressing 'z' button, zoom camera into selected object